cleaned up and refactor pollsWizard component
