+++ src/components/index.tsx
@@ -5,7 +5,7 @@ import { BoostIndicatorAnimation, PulseAnimation, SpinIndicator } from './animat
 import BeneficiaryModal from './beneficiaryModal/beneficiaryModal';
 import SnippetsModal from './snippetsModal/snippetsModal';
 import { UploadsGalleryModal } from './uploadsGalleryModal/container/uploadsGalleryModal';
-import { PollsWizardModal } from './pollsWizardModal';
+import { PollsWizardContent } from './pollsWizardContent';
 import { BottomTabBar } from './bottomTabBar';
 import { CheckBox } from './checkbox';
 import { CircularButton, TextButton, SquareButton } from './buttons';
@@ -207,7 +207,7 @@ export {
   SideMenu,
   SnippetsModal,
   UploadsGalleryModal,
-  PollsWizardModal,
+  PollsWizardContent,
   SpinGame,
   SpinIndicator,
   SquareButton,
