+++ src/components/pollsWizardContent/children/pollConfig.tsx
@@ -1,5 +1,5 @@
 
-import React, { forwardRef, useImperativeHandle, useRef, useState } from 'react';
+import React, { forwardRef, useImperativeHandle, useRef } from 'react';
 import { Text, View } from 'react-native';
 import styles from '../styles/pollConfig.styles';
 import { FormInput } from '../../formInput';
