+++ src/components/pollsWizardContent/styles/pollConfig.styles.ts
