+++ src/components/pollsWizardContent/styles/pollsWizardContent.styles.ts
