--- src/components/pollsWizardModal/index.ts
@@ -1,1 +1,0 @@
-export * from './container/pollsWizardModal'
